Refactor studyjob controller
